backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.1
cryptography==44.0.1
